FIX: Integer overflow in point-to-point average journey times causing excessively low journey times (and therefore excessively high revenues) to be recorded in some instances.

+++ simline.h
@@ -234,7 +234,7 @@ public:
 	* The table of point-to-point average speeds.
 	* @author jamespetts
 	*/
-	koordhashtable_tpl<id_pair, average_tpl<uint16> > * average_journey_times;
+	koordhashtable_tpl<id_pair, average_tpl<uint32> > * average_journey_times;
 
 };
